Import release tarballs for avahi.org/downloads
Import the old release tarballs to keep the existing download links
working.

It's unfortunately not possible to do a 301 redirect or similar with
GitHub Pages, so we can't redirect to binaries hosted on the GitHub
releases page.

It's not ideal to import a bunch of binaries to a git repo but I don't
see a good way to keep the existing links working otherwise.

Each release is <1.5MB and the total size is only 33MB so it seems
workable as a compromise.